BREAKING 2.0 CHANGES
first(), last(), firstKey(), lastKey(), sort() are now gone. Though
they may *seem* to work in some situations, they are not predictable
and may break or express unexpected behaviours.

_array, _keyArray() props removed. They were never used properly
in the first place, and will be eventually replaced with a single
_entries prop that will be properly updated when indexes roll out.

purge() is being removed, as level.destroy() is not available
in level-party now used by native-level-promise.

The "save" prop on set has been removed. I realized that was a
very stupid idea, either I need to redo the implementation or
simply use separate features NOT in this module.