fix: inconsistent jsdoc & methods
This PR fixes the inconsistency of the `#has`/`#delete` methods, which does not stringify the keys. If an array is passed, this should behave identically same to the `#get` method (which stringifies the key). Alongside, it also fixes some JSDoc inconsistency for the return values, and also changes the reference to the discord.js Collection (which was moved to an entirely different repository).